+++ CHANGELOG.md
@@ -5,6 +5,12 @@ branches do not necessarily reflect the changes that have gone into that branch.
 
 All notable changes to Sourcegraph are documented in this file.
 
+## Unreleased
+
+### Changed
+
+- Updating `maxReposToSearch` site config no longer requires a server restart to take effect.
+
 ## 2.12.0
 
 ### Changed