+++ CHANGELOG.md
@@ -7,6 +7,12 @@ All notable changes to Sourcegraph are documented in this file.
 
 ## Unreleased
 
+### Changed
+
+- Updating `maxReposToSearch` site config no longer requires a server restart to take effect.
+- We use GitHub's REST API to query affliated repositories. This API has wider
+  support on older GitHub enterprise versions.
+
 ### Removed
 
 - The experimental feature flag to disable the new repo update scheduler has been removed.